# specs/001-fullstack-web-todo/plan/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: REST API for the Todo application
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /tasks:
    get:
      summary: Get all tasks
      description: Retrieve a list of all tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Internal server error
    post:
      summary: Create a new task
      description: Create a new task with title and optional description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '422':
          description: Validation error
        '500':
          description: Internal server error
  /tasks/{id}:
    get:
      summary: Get a task by ID
      description: Retrieve a specific task by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
        '500':
          description: Internal server error
    put:
      summary: Update a task
      description: Update an existing task with new title and description
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
        '422':
          description: Validation error
        '500':
          description: Internal server error
    delete:
      summary: Delete a task
      description: Delete a specific task by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
        '500':
          description: Internal server error
  /tasks/{id}/toggle:
    patch:
      summary: Toggle task completion status
      description: Toggle the completion status of a specific task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
        '500':
          description: Internal server error
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - is_completed
        - created_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Title of the task
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          example: "Milk, eggs, bread"
        is_completed:
          type: boolean
          description: Completion status of the task
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2023-12-10T10:00:00Z"
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the task (required)
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          example: "Milk, eggs, bread"
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: New title of the task
          example: "Buy groceries and cook dinner"
        description:
          type: string
          description: New description of the task
          example: "Milk, eggs, bread, chicken"
        is_completed:
          type: boolean
          description: New completion status of the task
          example: true